refactor: move service logic to dedicated package and enhance service management functions
